+++ _data/series.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const { loadSeries } = require('../build/series');
+
+module.exports = async function () {
+  return loadSeries();
+};